Initilize CSS Styles
Made all styles for the initial page which is available to the user. The body was made into a flexbox so I can center the weather box. Other elements are set proportionally to the screensize, so no overlapping, or inline elements are possible. e.g. city input is 75% of the weather box which is 50% of the screeensize.