Fixed bugs
Essentially the fetch statements were synchronous and all of them ran regardless of errors in the first one. I move the last 2 fetches to a .then that only runs if the previous .then resolves it's promise. This allows data to remain there until the user input is successful. Also fixed the city not valid alert handeling. Next, I want to add more comments, give more console.logs for debugging purposes and add a dark mode.